refactor(phase8-b): domainize Phase 3 keep-celery tasks (D4a)
Per 符炫炜 D4a + PM Option A (msg=390d8c3f): mechanical move-only.
Move keep-celery tasks from `config/celery_tasks.py` (1523 LOC) into
domain-owned task modules. No behavior change. Task identity preserved
via explicit `name="config.celery_tasks.<orig>"` decorator argument.

New domain task modules:
- `aperag/domains/indexing/tasks.py` (1106 LOC):
  - 4 main tasks (parse/create/delete/update_index_task)
  - 4 orchestration tasks kept as-is (trigger_create/delete/update_indexes_workflow,
    notify_workflow_complete) — sync helper conversion deferred to task #40
  - 3 workflow entry-points (create/delete/update_document_indexes_workflow)
  - reconcile_indexes_task
  - BaseIndexTask + all index-related private helpers
- `aperag/domains/knowledge_base/tasks.py` (322 LOC):
  - reconcile_collection_summaries_task / collection_init_task /
    collection_delete_task / collection_summary_task / cleanup_expired_documents_task
  - Collection-summary-specific private helpers
  - export_collection_task NOT moved (defers to #39 ExportTask ORM carve;
    G1 boundary requires domain-local ORM mirror first)
- `aperag/domains/knowledge_graph/tasks.py` (28 LOC):
  - generate_graph_curation_run_task

`config/celery_tasks.py` trimmed to 109 LOC of evaluation legacy tasks
(scheduled for deletion in #39). `config/export_tasks.py` and
`config/celery.py` `include=[...]` updated to point at new modules.

Caller updates (10 sites):
- aperag/tasks/{scheduler.py,reconciler.py}
- aperag/domains/knowledge_base/service/{collection_service.py,document_service.py}
- aperag/graph_curation/service.py
- tests/unit_test/tasks/test_reconciler.py (6 monkeypatch path strings)
- tests/unit_test/graph_curation/test_service.py

Acceptance gates:
- pytest tests/unit_test/test_modularization_boundaries.py -x → 20 passed,
  0 failed (all G1-G19 green; export_collection_task deferred to keep G1
  domain-local-ORM ban intact)
- python -c "from config.celery import app; ..." → all 21 task names
  preserved (parse/create/delete/update_index_task,
  trigger_*_indexes_workflow, notify_workflow_complete,
  reconcile_indexes_task, reconcile_collection_summaries_task,
  collection_*_task, cleanup_expired_documents_task,
  generate_graph_curation_run_task, export_collection_task,
  reconcile_evaluations_task, initialize_evaluation_task,
  process_evaluation_batch_task, process_evaluation_item_task,
  run_evaluation_run)
- pytest tests/unit_test/tasks -x → 16 passed, 0 failed

Scope guards (per PM owner-subtree-only coordination):
- Did NOT touch aperag/concurrent_control/ (#38 owner)
- Did NOT touch aperag/db/, aperag/views/auth.py, evaluation_service.py (#36/#39 owners)
- Did NOT carve export_collection_task (waits on #39 ExportTask ORM carve)
- 4 trigger_* + notify_workflow_complete kept as Celery tasks (#40 scope)
- task name strings unchanged across all moves
- beat_schedule task identifiers unchanged

Combined with previous commit (D5 path moves) closes #37 D4a + D5.

Ghost-check: none.
